refactor: do atomic structure changes
QW-105
