+++ src/components/molecules/HomeListPoint/index.ts
@@ -1,0 +1,3 @@
+import HomeListPoint from "./HomeListPoint";
+
+export default HomeListPoint;
