+++ src/components/pages/ForgotPassword/ForgotPassword.tsx
@@ -23,13 +23,13 @@ const ForgotPassword = () => {
       email: !email.trim()
         ? EErrors.required
         : !emailPattern.test(email.trim())
-        ? "Введите корректный email"
+        ? EErrors.email
         : "",
       code: !code.trim() ? EErrors.required : "",
       password: !password.trim()
         ? EErrors.required
         : password.trim().length < 8
-        ? "Пароль должен содержать не менее 8 символов"
+        ? EErrors.password
         : "",
     };
 
@@ -42,7 +42,7 @@ const ForgotPassword = () => {
       navigate("Login", { direction: "backward" });
       showSuccessToast("Пароль сменен, зайдите с новыми данными");
     } else {
-      showErrorToast("Сначала заполните поля формы");
+      showErrorToast(EErrors.fields);
     }
   };
