+++ src/constants/errors.ts
@@ -1,3 +1,10 @@
 export enum EErrors {
   required = "Это обязательное поле",
+  email = "Введите корректный email",
+  password = "Пароль должен содержать не менее 8 символов",
+  inn = "Инн должен содержать 10 или 12 цифр",
+  fields = "Сначала заполните поля формы",
+  noChanges = "Вы не поменяли данные",
+  chooseDates = "Выберите обе даты",
+  timeDates = "Начальная дата не может быть позже конечной",
 }
