+++ src/hooks/useAccountStore.ts
@@ -1,7 +1,11 @@
 import { create } from "zustand";
 import { IStoreStatus } from "../model/misc";
 import { IAccount, initialAccounts } from "../constants/account";
-import { showErrorToast, showInfoToast, showSuccessToast } from "../helpers/toast";
+import {
+  showErrorToast,
+  showInfoToast,
+  showSuccessToast,
+} from "../helpers/toast";
 import { roles } from "../constants/roles";
 import { EErrors } from "../constants/errors";
 import { emailPattern } from "../constants/patterns";
@@ -73,12 +77,12 @@ const useAccountStore = create<IUseAccountStore>((set, get) => ({
       login: !login.trim()
         ? EErrors.required
         : !emailPattern.test(login.trim())
-        ? "Введите корректный email"
+        ? EErrors.email
         : "",
       password: !password.trim()
         ? EErrors.required
         : password.trim().length < 8
-        ? "Пароль должен содержать не менее 8 символов"
+        ? EErrors.password
         : "",
     };
 
@@ -113,7 +117,7 @@ const useAccountStore = create<IUseAccountStore>((set, get) => ({
         showErrorToast("Сначала корректно заполните поля формы");
       }
     } else {
-      showInfoToast("Вы не поменяли данные");
+      showInfoToast(EErrors.noChanges);
     }
   },
